flask-app/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.0
werkzeug>=0.15.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-ext-flask
opentelemetry-ext-requests
<<<<<<< HEAD
opentelemetry-exporter-cloud-monitoring # metrics export to Cloud Monitoring
opentelemetry-exporter-cloud-trace # trace export to Cloud Trace
=======
opentelemetry-ext-opencensusexporter # OC span + metrics export to OTLP collector
opentelemetry-exporter-cloud-monitoring # metrics export to Cloud Monitoring
opentelemetry-exporter-cloud-trace # trace export to Cloud Trace

# DEPRECATED: opentelemetry-ext-otcollector
# FUTURE OC EXPORTER: opentelemetry-ext-otlp
>>>>>>> agentless-setup
